Modify : mock data를 활용하여 여러 개의 피드 구현과 컴포넌트분리
